requirements.txt
Django==2.2
djangorestframework==3.8.2
idna==2.8
pytz==2019.2
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.6
